requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Machine Learning
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Validação e configuração
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Visualização
matplotlib==3.8.4
seaborn==0.13.2

# Cache / mensageria
redis==5.0.4

# Monitoramento
evidently==0.4.30
streamlit==1.35.0

# Testes
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Qualidade de código
black==24.4.2
flake8==7.0.0
